split the big script into modules
