+++ vectors_benchmark/datasets/dataset.json
@@ -4,7 +4,7 @@
       "name": "large",
       "shards": {
         "url": "https://storage.googleapis.com/nucliadb_indexer/4508ea2f-b3e2-41ca-89b6-224172f5bb3e.tar.gz",
-        "root_path": "shards",
+        "root_path": "data/shards",
         "ids": [
           "4508ea2f-b3e2-41ca-89b6-224172f5bb3e"
         ]
